Use `vulkaninfo` for AMD gpu heuristics on Windows
* Change exception paths for gpu heuristics function to return a
two part tuple like the non-exception paths, rather than a three
part one with `False`on the end. This prevents the caller from
blowing up with an extra exception.
* Added a path to get the AMD devices via vulkaninfo if we are on
windows as we will not have have rocminfo there even if the rocm
SDK is installed (windows rocm SDK :shrug:), and we don't
require the windows SDK anyway, so we can't use hipinfo either.
* The new code path, uses the existing AMD device/memory list to get
the memory info, and only returns devices that are on that list.