Add --checkforupdates argument, fix saves
If enabled, the argument --checkforupdates will fetch the Kobold-ROCm release page one time on start up via HTTPS and compare the latest version number with the current version number and notify the user if a new version is available.
A GUI button is shown on the Network tab.
Fixed a bug where importing a save file wouldn't choose the hipBLAS option